requirements.txt
aniso8601==6.0.0
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.4.1
Flask==1.0.2
Flask-RESTful==0.3.7
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
more-itertools==7.0.0
mysql-connector-python==8.0.15
pluggy==0.9.0
protobuf==3.7.1
py==1.8.0
pytest==4.4.1
pytz==2019.1
redis==3.3.11
requests==2.22.0
six==1.12.0
urllib3==1.24.2
Werkzeug==0.15.2
